Update github actions to include more triggers + go versions; Update to io.util
