+++ .github/workflows/unit_test.yml
@@ -6,7 +6,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        go-version: ['1.18.x']
+        go-version: ['1.18.x', '1.19.x', '1.20.x', '1.21.x', '1.22.x', '1.23.x']
 
     steps:
       - uses: actions/checkout@v3
@@ -15,7 +15,9 @@ jobs:
         uses: actions/setup-go@v3
         with:
           go-version: ${{ matrix.go-version }}
-          cache: true
+          # As there is only one Golang script being run.
+          # Faster if we don't cache as the cache is per commit anyway.
+          cache: false
 
       - name: Golang version
         run: go version
